+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Allow escape to close the exit modal
 - Allow enter to exit the app from the exit modal
 - Add numbered hotkeys to the environment selection modal
-- Update readme
+- Update the about section of the readme
+- Refactor how modals and their background content is managed
 
 # v3.3.0